[#8] Added tabItem, subTabItem; moved enums to model package
