+++ server/src/main/kotlin/com/sourceplusplus/portal/server/model/TraceType.kt
@@ -1,0 +1,10 @@
+package com.sourceplusplus.portal.server.model
+
+enum class TraceType {
+    LATEST,
+    SLOWEST,
+    FAILED;
+
+    val id = name.toLowerCase()
+    val description = name.toLowerCase().capitalize()
+}
