+++ server/src/main/kotlin/com/sourceplusplus/portal/server/template/Menu.kt
@@ -1,5 +1,6 @@
 package com.sourceplusplus.portal.server.template
 
+import com.sourceplusplus.portal.server.model.*
 import kotlinx.html.*
 
 fun FlowContent.menu(cssClasses: String = "ui accordion displaynone",
@@ -16,62 +17,8 @@ fun FlowContent.menu(cssClasses: String = "ui accordion displaynone",
     }
 }
 
-fun FlowContent.menuItem(pageType: PageType, isActive: Boolean = false, block: FlowContent.() -> Unit) {
-    pageType.applyThis(this, isActive, block)
-}
-
-enum class PageType {
-    OVERVIEW {
-        override fun applyThis(flowContent: FlowContent, isActive: Boolean, block: FlowContent.() -> Unit) {
-            flowContent.apply {
-                if (isActive) {
-                    a(classes = "item active_tab") { +"Dashboard" }
-                } else {
-                    a(classes = "item inactive_tab") {
-                        id = "sidebar_overview_link"
-                        href = "overview"
-                        +"Dashboard"
-                    }
-                }
-            }
-        }
-    },
-    TRACES {
-        override fun applyThis(flowContent: FlowContent, isActive: Boolean, block: FlowContent.() -> Unit) {
-            flowContent.apply {
-                var activeClass = "active_tab"
-                if (!isActive) {
-                    activeClass = "inactive_tab"
-                }
-                div("title item") {
-                    classes = classes.plus(activeClass)
-                    i("dropdown icon") {
-                    }
-                    +"Traces"
-                }
-                div("content") {
-                    block()
-                }
-            }
-        }
-    },
-    CONFIGURATION {
-        override fun applyThis(flowContent: FlowContent, isActive: Boolean, block: FlowContent.() -> Unit) {
-            flowContent.apply {
-                if (isActive) {
-                    a(classes = "item active_tab") { + "Configuration" }
-                } else {
-                    a(classes = "item inactive_tab") {
-                        id = "sidebar_configuration_link"
-                        href = "configuration"
-                        +"Configuration"
-                    }
-                }
-            }
-        }
-    };
-
-    abstract fun applyThis(flowContent: FlowContent, isActive: Boolean = false, block: FlowContent.() -> Unit)
+fun FlowContent.menuItem(pageType: PageType, isActive: Boolean = false, block: FlowContent.(activeClass: String) -> Unit) {
+    pageType.applyThis(this, true, isActive, block)
 }
 
 fun FlowContent.subMenuItem(vararg traceTypes: TraceType = arrayOf()) {
