+++ server/src/main/kotlin/com/sourceplusplus/portal/server/template/Navbar.kt
@@ -1,5 +1,6 @@
 package com.sourceplusplus.portal.server.template
 
+import com.sourceplusplus.portal.server.model.TimeIntervalType
 import kotlinx.html.*
 
 fun FlowContent.overviewNavBar(block: FlowContent.() -> Unit) {
@@ -141,12 +142,4 @@ fun FlowContent.leftAlign(block: FlowContent.() -> Unit) {
 
 fun FlowContent.rightAlign(block: FlowContent.() -> Unit) {
     block()
-}
-
-enum class TimeIntervalType(val id: String, val description: String) {
-    FIVE_MINUTES("5_minutes", "5 MINUTES"),
-    FIFTEEN_MINUTES("15_minutes", "15 MINUTES"),
-    THIRTY_MINUTES("30_minutes", "30 MINUTES"),
-    ONE_HOUR("hour", "HOUR"),
-    THREE_HOURS("3_hours", "3 HOURS");
 }
