+++ server/src/main/kotlin/com/sourceplusplus/portal/server/template/Tab.kt
@@ -1,52 +1,17 @@
 package com.sourceplusplus.portal.server.template
 
+import com.sourceplusplus.portal.server.model.*
 import kotlinx.html.*
 
 fun FlowContent.tabs(block: FlowContent.() -> Unit) {
     block()
 }
 
-fun FlowContent.overviewTab(isActive: Boolean = false) {
-    if (isActive) {
-        a(classes = "ui dropdown item active_tab") {
-            i("icon demo-icon dashboard") {
-            }
-        }
-    } else {
-        a(classes = "ui item hide_on_toggle") {
-            id = "overview_link"
-            href = "overview"
-            i("icon demo-icon dashboard inactive_tab") {
-            }
-        }
-    }
+fun FlowContent.tabItem(pageType: PageType, isActive: Boolean = false, block: FlowContent.(activeClass: String) -> Unit) {
+    pageType.applyThis(this, false, isActive, block)
 }
 
-fun FlowContent.tracesTab(isActive: Boolean = false, block: FlowContent.(activeClass: String) -> Unit) {
-    var activeClass = "active_tab"
-    if (!isActive) {
-        activeClass = "inactive_tab"
-    }
-    block(activeClass)
-}
-
-fun FlowContent.configurationTab(isActive: Boolean = false) {
-    if (isActive) {
-        a(classes = "ui dropdown item active_tab") {
-            i("icon configure") {
-            }
-        }
-    } else {
-        a(classes = "ui item hide_on_toggle") {
-            id = "configuration_link"
-            href = "configuration"
-            i("icon configure inactive_tab") {
-            }
-        }
-    }
-}
-
-fun FlowContent.traces(activeClass: String = "", vararg traceTypes: TraceType = arrayOf()) {
+fun FlowContent.subTabItem(activeClass: String = "", vararg traceTypes: TraceType = arrayOf()) {
     div("ui dropdown item") {
         classes = classes.plus(activeClass)
         unsafe {
@@ -64,13 +29,4 @@ fun FlowContent.traces(activeClass: String = "", vararg traceTypes: TraceType = 
             }
         }
     }
-}
-
-enum class TraceType {
-    LATEST,
-    SLOWEST,
-    FAILED;
-
-    val id = name.toLowerCase()
-    val description = name.toLowerCase().capitalize()
 }
